Add Omnes to use it for pub/sub
Using Omnes will be the default way to make usage of pub/sub features
from Solidus 3.2 onwards.

This commit prepares the stage by adding it as a dependency. A new
`Spree::Config#use_legacy_events` preference has been added to allow
using the old system based on `ActiveSupport::Notifications`. We still
keep it as the default system because internal subscribers still only
work with it.